Achievements: 

1. Retrieve a collection of metadata attributes of each individual object all at once (with PDC, each single metadata attribute requires a request)
2. Retrieve a collection of metadata attributes from the sub-objects of major object all at once (with PDC, there is no concept of compound object)
3. Retrieve a slice of data from any single individual object associated with an NDArray (theoretically any number of dimensions)
4. Retrieve a collection of data slices from all or specified set of sub-objects all at once (with PDC, loading each variable requires a single request)

In-Progress: 
 
1. Integrating to HydraGNN and preparing the benchmark on Perlmutter (in progress)
	a. The first test I would like to compare is the effect of having compound data objects and aggregated data fetching requests. I plan to get the throughput/latency of aggregated requests and unaggregated requests, for both data and metadata. The overall training speedup with aggregated mode.

2.  I need to add probes to the requests to include information on training instance ID and data loader rank ID, so that I can analyze the data loading request distribution among all data servers. 

Plans:

1. To collect the benchmarking results and plot them
2. To analyze the request distribution when parallel training happens.

+++ server/src/srvctx/mod.rs
@@ -81,11 +81,6 @@ impl ServerContext {
         let (shutdown_tx, shutdown_rx) = oneshot::channel();
         let _ = endpoint.listen(start_listen_tx, shutdown_rx).await?;
         debug!("Endpoint {} is listening now", id);
-        let ready_file = FileUtility::get_pdc_tmp_dir().join(format!("rx_{}_ready.txt", id));
-        // write 'ready' to the ready file
-        tokio::fs::write(&ready_file, "ready")
-            .await
-            .expect("Failed to create ready file");
         let endpoint = Arc::new(TokioMutex::new(endpoint));
         self.endpoints.insert(id.to_string(), endpoint);
         self.endpoint_shutdowns.insert(id.to_string(), shutdown_tx);
@@ -227,19 +222,24 @@ impl ServerContext {
             if let Err(_) = tx.send(()) {
                 warn!("Failed to send shutdown signal to endpoint: {}", id);
             }
-            let ready_file = FileUtility::get_pdc_tmp_dir().join(format!("rx_{}_ready.txt", id));
-            // delete the ready file
-            if let Err(e) = tokio::fs::remove_file(&ready_file).await {
-                warn!("Failed to remove ready file for endpoint {}: {}", id, e);
-            }
 
-            let server_list_path = FileUtility::get_pdc_tmp_dir().join(format!("rx_{}.txt", id));
-            // delete the server list file
-            if let Err(e) = tokio::fs::remove_file(&server_list_path).await {
-                warn!(
-                    "Failed to remove server list file for endpoint {}: {}",
-                    id, e
-                );
+            if self.rank == 0 {
+                let ready_file =
+                    FileUtility::get_pdc_tmp_dir().join(format!("rx_{}_ready.txt", id));
+                // delete the ready file
+                if let Err(e) = tokio::fs::remove_file(&ready_file).await {
+                    warn!("Failed to remove ready file for endpoint {}: {}", id, e);
+                }
+
+                let server_list_path =
+                    FileUtility::get_pdc_tmp_dir().join(format!("rx_{}.txt", id));
+                // delete the server list file
+                if let Err(e) = tokio::fs::remove_file(&server_list_path).await {
+                    warn!(
+                        "Failed to remove server list file for endpoint {}: {}",
+                        id, e
+                    );
+                }
             }
         }
